helpers: remove TemplateField helper method
This was the only method currently using the html/template package.  The
Tailscale mobile applications don't allow use of html/template to
conserve memory usage. This prevented us from including the new web
interface in the 1.56 releases on mobile. We weren't using this helper
method anyway, so removing it should allow us to include the web client
in Tailscale mobile apps.